+++ index.html
@@ -24,7 +24,7 @@
         </form>
         <div id="books-container" class="d-flex flex-wrap"></div>
     </div>
-    <script type="module" src='./js/dom.js'></script>
-    <script type="module" src='./js/app.js'></script>
+    
+    <script  src='./js/app.js'></script>
 </body>
 </html>
